+++ plugins/comparative-adapters/src/PAFAdapter/util.test.ts
@@ -1,0 +1,7 @@
+import { flipCigar } from './util'
+
+test('flip cigar', () => {
+  expect(flipCigar(['3', 'M', '5', 'D', '5', 'M', '5', 'I', '6', 'M'])).toEqual(
+    ['6', 'M', '5', 'I', '5', 'M', '5', 'D', '3', 'M'],
+  )
+})
